TODO: why all_types.dart slow for flutter
I really had to stop here so yeah a commit msg as a TODO.

Next: update the CLI to use TypeGraph (or maybe get rid of that and let the Engine generate the graph)

The slow part with flutter (~2400 types/subtypes) is the `transitiveReduction` code which is inaccurate anyway.

Maybe there is a better way to construct the graph with minimal edges (without transitive ones).

Another TODO: logging.. I added logging quckly and it's manual. It was to discover where the code was slow for the flutter repo case.